requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
redis==5.0.4
pydantic==2.7.4
pydantic-settings==2.3.1
python-json-logger==2.0.7
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-logging==0.47b0
opentelemetry-instrumentation-redis==0.47b0
httpx>=0.25.0
pyyaml>=6.0

